Refactoring functions

+++ section_7/mysql/functions.php
@@ -1,6 +1,30 @@
 <?php include "db.php"; ?>
 <?php
 
+// Create records in the users database
+function createRows() {
+
+    global $connection;
+
+    if(isset($_POST['submit'])) {
+
+        $username = $_POST['username'];
+        $password = $_POST['password'];
+
+        $query = "INSERT INTO users(username,password) ";
+        $query .= " VALUES ('$username', '$password')"; 
+
+        $result = mysqli_query($connection, $query);
+        if(!$result) {
+            die("QUERY FAILED" . mysqli_error());
+        } else {
+            echo "Record Created";
+        }
+
+    }   
+
+}
+
 // Show ids in the select dropdown
 function showAllData() {
 
@@ -24,18 +48,23 @@ function updateTable() {
 
     global $connection;
 
-    $username = $_POST['username'];
-    $password = $_POST['password'];
-    $id = $_POST['id'];
+    if(isset($_POST['submit'])) {
 
-    $query = "UPDATE users SET ";
-    $query .= "username = '$username', ";
-    $query .= "password = '$password' ";
-    $query .= "WHERE id = $id ";
+        $username = $_POST['username'];
+        $password = $_POST['password'];
+        $id = $_POST['id'];
 
-    $result = mysqli_query($connection, $query);
-    if(!$result) {
-        die("QUERY FAILED" . mysqli_error($connection));
+        $query = "UPDATE users SET ";
+        $query .= "username = '$username', ";
+        $query .= "password = '$password' ";
+        $query .= "WHERE id = $id ";
+
+        $result = mysqli_query($connection, $query);
+        if(!$result) {
+            die("QUERY FAILED" . mysqli_error($connection));
+        } else {
+            echo "Record Updated";
+        }
     }
 
 }  
@@ -43,18 +72,24 @@ function updateTable() {
 // Delete records in the database
 function deleteRows() {
 
-    global $connection;
+    if(isset($_POST['submit'])) {
 
-    $username = $_POST['username'];
-    $password = $_POST['password'];
-    $id = $_POST['id'];
+        global $connection;
 
-    $query = "DELETE FROM users ";
-    $query .= "WHERE id = $id";
+        $username = $_POST['username'];
+        $password = $_POST['password'];
+        $id = $_POST['id'];
 
-    $result = mysqli_query($connection, $query);
-    if(!$result) {
-        die("QUERY FAILED" . mysqli_error($connection));
+        $query = "DELETE FROM users ";
+        $query .= "WHERE id = $id";
+
+        $result = mysqli_query($connection, $query);
+        if(!$result) {
+            die("QUERY FAILED" . mysqli_error($connection));
+        } else {
+            echo "Record Deleted";
+        }
+
     }
 
 }
